Add a utility class to the testing infrastructure
This adds a “zip” function which allows the elements of multiple arrays to be joined into a single array.

This adds a cross product iterator. There are two modes. The first mode allows you to use a single array repeated a given number of times. In the second mode, if the repeat count is set to zero, then it is assumed that all elements of the array are arrays and each will be iterated in sequence.